+++ Makefile
@@ -1,0 +1,2 @@
+all:
+	python eth2_testgen/shuffling/tgen_shuffling.py test_vectors/test_vector_shuffling.yml
