+++ eth2_testgen/shuffling/constants.py
@@ -1,0 +1,4 @@
+
+SHARD_COUNT = 2**10  # 1024
+EPOCH_LENGTH = 2**6  # 64 slots, 6.4 minutes
+TARGET_COMMITTEE_SIZE = 2**8  # 256 validators
