+++ requirements.txt
@@ -1,0 +1,1 @@
+eth-typing==2.0.0
